+++ CHANGELOG.md
@@ -12,6 +12,11 @@ captured by the sniffer
 - Add `-o/--output` option which give the user the possibility to save the
 output from sniffed packets to a file for later reviewing
 - Improve the UI by displaying the HTTP packets formatted properly and stylish
+  - The packets will fit in a panel and only take the required space to be
+displayed, nothing more, nothing less
+
+### Changed
+- Rename the footer bar from **Information** to **Status**
 
 ## [0.4.0] - 2021-12-28
 ### Added